Rearrange functions to avoid circular dependency
