+++ packages/contractkit/src/identity/claims/account.test.ts
@@ -5,7 +5,8 @@ import { NativeSigner } from '@celo/utils/lib/signatureUtils'
 import { newKitFromWeb3 } from '../../kit'
 import { IdentityMetadataWrapper } from '../metadata'
 import { createAccountClaim, MetadataURLGetter } from './account'
-import { Claim, verifyClaim } from './claim'
+import { Claim } from './claim'
+import { verifyClaim } from './verify'
 
 testWithGanache('Account claims', (web3) => {
   const kit = newKitFromWeb3(web3)
